Abolish differentiation between SegmentedSentence and TaggedSentence
Code show that in all cases there two versions of the same method, one of which deals with TaggedSentence and the other deals with SegmentedSentence. Profiling showed that most time is not consumed here, therefore there's no need to make the code more complex and difficult to understand just to save a very little amount of time. Thus, the method which deals with SegmentedSentence is deleted and in Thulac.java, the segments are extracted from the list of WordWithTag.