Add a test for this program
Junit is used for testing and added to dependencies of the test.
This test is mainly used to automate running process and output speed and accuracy. The output directory is tmp/output.txt, input is src/test/resources/data_input.txt and the right answer for comparison is located at src/test/resources/data_seg.txt. However, these 3 files are not added to this repo and thus not versioned, for the simple reason that they are too big.
The test files are from https://github.com/isnowfy/snownlp, under the MIT license.