Use interfaces to standardize I/O
The main() entrance is moved to Main.java, Thulac.java is now clean of I/O details.
The interfaces IInputProvider and IOutputHandler are used to standardize I/O operations, see their documentation for more information.
This allows users to deal with raw data provided by the segmentation program, in form of List<TaggedWord>, instead of printed String. For example, users might want to send the words directly through the internet without having to use a temp file. Additionally, this also leaves handling of charsets and other I/O details to the user (or simply use the static methods int IInputProvider.java and IOutputHandler.java), saving the effort of having to transfer GB2312 to UTF-8 at the beginning.